requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
py-cord==2.4.1
pycparser==2.21
PyNaCl==1.5.0
pyTelegramBotAPI==4.13.0
requests==2.31.0
urllib3==2.0.4
websocket-client==1.6.2
yarl==1.9.2
